+++ src/structure/ray.rs
@@ -11,7 +11,7 @@ impl Ray {
         Self { origin, direction }
     }
 
-    pub fn position(&self, t: f32) -> Point {
+    pub fn position(&self, t: f64) -> Point {
         self.origin.clone() + self.direction.clone() * t
     }
 }
